+++ R/esp_get_nuts.R
@@ -91,7 +91,8 @@
 #' geographic position of the Canary Islands. When using the output for
 #' spatial analysis or using tiles (e.g. with [esp_getTiles()] or
 #' [addProviderEspTiles()])  this option should be set to `FALSE` in order to
-#' get the actual coordinates, instead of the modified ones.
+#' get the actual coordinates, instead of the modified ones. See also
+#' [esp_move_can()] for displacing stand-alone \CRANpkg{sf} objects.
 #'
 #' @examples
 #'